server(index.js) Add general Mutations and Queries to server + @auth directive.
Created several basic Mutations, Subscriptions and Queries to manage creation and modification of data in database.

Created @auth directive to manage permissions in graphql. This will allow for Graphql level permissioning + the authenticated layer to ensure client is logged in.

**Breaking changes.
Substituted as well apollo-server-express for apollo-server. Server won't depend any more of express.