server(index.js) Add Sequelize connection through AWS Aurora + Folio functionality + AWS S3 client + Models updates + Graphql basic queries.
New Sequelize client configured to connect to AWS Aurora CLI to upload Tickets as long term registry. Added sequelize and mysql2 libraries for support.

*** Breaking changes!
.env was updated due to new AWS credentials to connect! It is important to update server .env to be able to work + install necessary dependencies.

Folio functionality was solved by adding a new Model to the MongoDB that keeps track of A and B folios' count for the tickets.

No breaking changes!

AWS S3 Client support file and image(base64) uploads to our S3 service.

*** Breaking changes
New .env variables were added to have AWS credentials secured to connect to the service. In addition to this, whenever a new IP address requires to connect to AWS services, it is necessary to whitelist it inside aws.amazon.com

MongoDB models were updated to fix emerging conflicts to allow for the server to support new requirements. This of course included some additional GraphQL queries for the server to handle file and image uploads, Tickets initialization and process handling, Subscriptions to support live reloading inside the application, etc.

*** Breaking changes!!
All content in DB must be dropped and tables have to be recreated to support new Schemas.