feature ( index.js ) Create Turn functionality to track Turns inside the application.
Added new Turn Model to database to handle turns inside the application. A turn maps all of the tickets that happen within the time frame and are completed ( including outTruckImage and weighing ). If a Ticket is not completed, it goes on to the next turn.

Breaking changes include the new Turn model, forcing the backend to be updated.