feature ( index.js ) Add subscription to handle active tickets.
dashboard-tickets now subscribes to ticket updates depending on whether they're included in certain Turn or not. Some logic refactor was done in order to achieve this.

Previously, everytime a new Turn was created, all pending Tickets were automatically added to that specific Turn, nonetheless, this is an issue if Tickets that were already appended to a certain Turn weren't finished. Now, that Turn-Tickets relationship got extrapolated and is independent from whether there is a turn or not.

BREAKING CHANGES: This required some refactoring which could possible throw some bugs in the future!!!