feature ( index.js ) New queue view to place done Tickets for a specific Turn.
Since backend got some refactoring in terms of the Turn-Ticket mapping, a view refactor of dashboard-tickets was necessary in order to visually update the state of the Tickets that were actually going to be included on a specific Turn. A new side list below the clock displays ALL tickets that will be included whenever this Turn ends.

No breaking changes.