bugfix ( index.js ) Subscriptions were not being properly unsubscribed.
Whenever a user logged into the application and navigated towards a dashboard that included Subscriptions and then unmounted the component, the Subscriptions were not being removed as well, causing an overflow of subscriptions, reaching the EventEmitter.defaultMaxListeners ( 10 ). Fixed the issue by conditionally creating a websocket if it was nonexistent and increased EventEmitter.defaultMaxListeners to 30 since there are several subscriptions happening.

No breaking changes. This solutions has yet to prove its effectiveness in production conditions.