bugfix ( index.js ) Remove array passing statement from pubsub asyncIterator.
According to PubSub documentation it is possible to pass an array or a sole string to the asyncIterator function to build the subscription channels.
Currently an array is being passed and it is probably causing a cyclical behavior on the subscriptions. This string passing is a test to evaluate its
behavior.

No breaking changes.