hotfix ( index.js ) Refactor REGISTER_TRUCK_INIT functionality inside GUARD dashboard-registry view
Beforehand, image upload to AWS S3 implementation happened separately from initiating a Ticket inside the application. Since it was not possible to know whether the image was properly uploaded or not, it was necessary to split the operations.

Now, an implementation of transactions using mongoose-transactions was achieved, allowing both operations ( uploading file to S3 and registering the truck ) can be done on the same mutation. Thanks to the Transaction property of atomicity, either all the operations are successfully implemented, or none at all.

No breaking changes.