feature(graphql-client): Refactor graphql client
Graphql client was configured not so well, sending the cookie on headers, which is not a best practice. Also, an error handler was set in order to get the stacktrace of every graphql error in the app.

There are still some needed fixes on backend. This commit breaks the app.