fix(auth): Get rid of cookies.
Apparently, there is no way to set a cookie and sent it cross-origin, so, we have a problem where the cookie cannot be set on browser without being ignored and requesting to the api, and if the cookie is set with the server domain, is ignored by the browser. Now they're sent over headers.

No breaking changes