feature (index.js) Create Reports section to support Reports generation
Created new Reports section to handle any type of reports that may be required to create in the future.
This section is intended to showcase a brief summary of sales and performance in terms of different object. The global view manages some global filters that apply to all reports. Once a report type is selected, it will display its specific filters, to further narrow down a search.

An integration with rechartsjs library was required to display the data in responsive charts. Each chart requires its own query to display data appropiately.

No breaking changes.