hotfix ( index.js ) General fixing of issues and typos in app
Did a whole lookup of small issuesm bugs, and typos to fix.

Breaking changes:

Beforehand, whenever a turn ended, all the tickets associated with the Turn were deleted from the MongoDB database and transferred to the AuroraDB. Since Tickets' data will be required to generate reports with results, it was decided to keep the same schema of migration to the AuroraDB, but persisting the data inside MongoDB as well. This, to persist the data and be able to do analysis with it. Using mongoose-delete to add a deletion parameter to all Tickets.