feature ( index.js ) Add LOADER view New LOADER view to handle product loading functionality at the loader checkpoint.
LOADER users have access to the active trucks pending product, and already served trucks. This view as well, offers the ability to update product type on a Ticket prior to its billing at the CASHIER view, this way, a Ticket is able to modify its product without having to cancel the Ticket.

No breaking changes.

Signed-off-by: oscardavidrm <A01229551@itesm.mx>