feature ( index.js ) Ticket attributes refactor to handle checkpoints and metadata
Ticket model had only a date attribute that was added on its creation. A need to increase precision in the states of a Ticket lifetime ( registering, loading, billing ) required a refactor of the date timestamp into three separate attributes: in, out, load.

This timestamp refactor allows better precision during a Ticket analysis. Due to Ticket audit requirements, a usersInvolved object was added to the Tickets to keep track of all the users that interacted with that Ticket.

*** Breaking changes
Replaced date attribute from Tickets and updated all resolvers and functions that referenced the attribute.

Signed-off-by: oscardavidrm <A01229551@itesm.mx>