Refactor Truck registry QR codes download type
Currently, inside registry, each Truck has its own QR code and whenever the user wishes to download it, the browser previews it and offers it to download it as pdf.

This behaviour is being refactored to download it as a png image type to ease the process of downloading the QR codes.

No breaking changes.

Signed-off-by: oscardavidrm <A01229551@itesm.mx>