fix(rock-price-requests): Add floor price and price as required to rock-price-requests
This is to have consitency in db. If only one is allowed, it will cause some problems when the actual price is changed for any reason, and the new price is lower than the floor price or vice versa. It was decided that the price request should have both prices to apply them as a whole. This guarantee that it won't cause inconsistencies.

No breaking changes