feature(cancel-tickets): Add ticket cancelation.
Ticket cancellation is enabled by mongoose-disabled library, which allows to hide the ticket from active tickets list. Subscriptions and queries were all updated with the disabled flag, and tickets can only be recovered within 24 hours from its cancelation. Also, a validation to not recover a ticket from a user that has another active ticket was added, and the ticket creation was updated because it validates to not have another.

This could break some subscriptions logic beacuse of the new flag. It is advised that probably will be situations were canceled tickets are shown due to mismanagement of disabled flag. We should be focused on checking this possible issue.