Refactor Client address to support a more complex type
Reformatted client address type to a new object with more attributes to support a deeper complexity no each client's address. Had to perform a backend and frontend refactor on all address references for Clients.

Breaking changes.
Due to the nature of a refactor, it is possible to experience new bugs that must be resolved in future commits.

Signed-off-by: oscardavidrm <A01229551@itesm.mx>