Feature (index.js) Create new Bills view to handle Ticket grouping
New Bills view adds the functionality of handling pending Tickets to be grouped and billed into the accounting system. The integration with the accounting system is pending and will be included in further functionality, therefore, this view will work as a canvas to continue over with the integration.

*** Breaking changes ***
Added new isBilled attribute to Tickets to keep track of which Tickets have been already billed, and should not be considered again.

Signed-off-by: oscardavidrm <A01229551@itesm.mx>