Hotfix (index.js) Fix clientEdit resolver
clientEdit resolver was not updated on new Client migration, so it was presenting some issues. Had to refactor and add balance attribute in the query to get Clients data.
Removed as well in client-special-price the top constraint to ensure that a price modification was kept between floorPrice and price of a Product. Due to business requirements, the top constraint had to be removed, allowing higher special prices than a base price of a product.

 *** No breaking changes. ***

Signed-off-by: oscardavidrm <A01229551@itesm.mx>