feature(unlimited-credit): Add unlimited credit
Unlimited credit is only an infinite credit limite, which in practice is impossible to reach. It was decided tu use this criteria in order to not have duplicate fields on database representing the same thing, this will avoid inconsistencies in the future. Also, interfaces were updated to handle it.

No breaking changes