Hotfixes (n-clients-per-truck, defaultCreditDays, Client Stores)
n-clients-per-truck
Refactored Truck plates uniqueness to support more than one client per Truck due to business logic. It is possible that several clients share the same truck, therefore, it was necessary to allow plates repetition. A front end logic in the access view allows us to handle situations of plates that are shared, allowing the GUARD to select between the possible options.

defaultCreditDays
Added defaultCreditDays attribute to Client object to support a default limit of days to pay for a credit. Added in new-client-form and edit-client-form the required input. The default value is set to 0;

 Client Stores
 As per business logic requested, a new Store object was created to handle the concept of stores for each client. The Store object affect Clients and Tickets.
 In the following commits, inside the tickets view, the Tickets that are from clients with available stores, should have the option to set the store.

Signed-off-by: oscardavidrm <A01229551@itesm.mx>