feature ( machine, machine-diesel-load, tank-diesel-load ) Add new Machine object and keep track of its Diesel loading history
machine.graphql
New machine object is intended to represent the working machines the business uses to work on a daily basis.

machine-diesel-load.graphql
The MachineDieselLoad object keeps track of all the diesel loads a given machine does. As per business requirements, it is necessary to keep track of these loads to evaluate truck performance in general.

tank-diesel-load.graphql
The TankDieselLoad object registers all diesel inputs that happen to the tank that the business uses to load diesel to its machines.

***No breaking changes.***

Signed-off-by: oscardavidrm <A01229551@itesm.mx>