feature ( Bills view level up and refactor ) Refactor and boost up of Bills view
Created new Bill object to handle storing of generated grouped bills that are constructed in the grouping view inside bills.

Refactored queries to be able to group bills and select required data to construct a complete bill which then can be transferred to the accounting system once it's integrated.

Added as well a bill history view to analyse archived bills past its creation.

*** Breaking changes ***
Since Bills as well require an independent folio, a new Folio object with name: F and count: 1000 must be created in the production environment to be able to create folios for generated bills.

Signed-off-by: oscardavidrm <A01229551@itesm.mx>