hotfix ( history ) Fix issues with History filters.
Product filter was sending the product ID rather than the product name, therefore it was not displaying anything.

Fixed as well the preset date filters for 'De ayer', 'De hoy' to hold a 1 day range to be able to query correctly.

Remove ticket-list folder since it was not being used for anything.

*** No breaking changes ***

Signed-off-by: oscardavidrm <A01229551@itesm.mx>