feature ( ClientsGroup view ) Create view to create and edit ClientsGroup.
Created new section to CRUD ClientsGroups inside the application. New groups can now be referenced when creating a new promotion to specify a pre built group of clients to apply a promotion at their level.

Added groups attribute to Promotion object to hold referenced groups so clients that belong to an included group, get the promotion available as well.

*** No breaking changes ***

Signed-off-by: oscardavidrm <A01229551@itesm.mx>