hotfix ( Cyclic relation in Weight effects ) Fix cyclic relation in ticket-submit-form that causes memory heap to fill.
In ticket-submit form there are two ways a user can enter the weight to submit; can either be manually or the scale automatically injects the data. And whenever the data is input by the user, it is necessary to apply the dynamic tariff behind the scenes, so the calculation must happen on the fly. New implementation avoids cyclic relations and only applies dynamic tariff when data is manually written, but when the scale is connected, it applies the dt on the go.

*** No breaking changes ***

Signed-off-by: oscardavidrm <A01229551@itesm.mx>