Hotfix ( Roles ) Add new requested roles as per business requirement
Created new roles such as COLLECTOR, COLLECTOR_AUX, SALES, TREASURER, and AUDITOR to assign different required roles. Implementation of new roles was a merging of existing roles and permissions, thus, it is possible to find inconsistencies and they should be addressed as soon as they're identified.

Current version verifies that the app didn't break, but some testing should happen to verify that everything is running smoothly.

*** No breaking changes ***

Signed-off-by: oscardavidrm <oscardavidrodriguezm@gmail.com>